set up cloud config module

+++ ExcelService/src/main/java/com/antra/evaluation/reporting_system/listener/ExcelCrudQueueListener.java
@@ -30,7 +30,11 @@ public class ExcelCrudQueueListener {
             throw new CrudTypeNotFoundException();
         } else if (request.getCrudType().equals("DELETE")) {
             String fileId = request.getExcelFileId();
-            excelService.deleteFile(fileId);
+            if (fileId != null) {
+                excelService.deleteFile(fileId);
+            } else {
+                throw new FileNotFoundException();
+            }
             log.info("File with fileId: {} has been deleted!", fileId);
         }
     }
